Generate_Competitors/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-sambanova
pydantic
langgraph
tavily-python
ipython
